File which is currently running changes color again. 	Fixed bugs with sorting and the color 	Fixed bugs with sorting causing the program to crash if it was done after running once. 	Fixed bugs with repainting code tabs being glitchy. Fixed not saving divider location. Got rid of extra spaces in output. Fixed missing dependencies in extensions and implementations. Made errors log to System.err and a text file so I can stop forgetting to enable the error file in releases and every commit.

+++ src/DropboxGrader/Config.java
@@ -43,7 +43,7 @@ public class Config {
         //GradingPanel
         autoRun=false;
         runTimes=1;
-        dividerLocation=Toolkit.getDefaultToolkit().getScreenSize().width*2/3; //2/3rds
+        dividerLocation=(int)(Toolkit.getDefaultToolkit().getScreenSize().width*2.0/3); //2/3rds
     }
     public static void readConfig(){
         if(!CONFIGFILE.exists()){
